finished making unit tests for posts and comments
